# docs/api/openapi-drf.yaml
openapi: 3.0.3
info:
  title: diconai API
  version: 1.0.0
  description: 산재 예방 통합 관제 시스템 — drf-server REST API
paths:
...
  /api/admin/geofences/:
    get:
      operationId: admin_geofences_retrieve
      description: 검색 + 필터. `/api/geofences/` ViewSet의 어드민 버전.
      summary: (관리자) 지오펜스 목록
      parameters:
      - in: query
        name: facility_id
        schema:
          type: integer
      - in: query
        name: name
        schema:
          type: string
        description: 이름 부분검색
      - in: query
        name: risk_level
        schema:
          type: string
        description: normal/warning/danger
      tags:
      - Admin — Geofence
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoFenceAdminListResponse'
          description: ''
    post:
      operationId: admin_geofences_create
      description: |-
        관리자 지오펜스 목록 조회 / 등록
        GET  /api/admin/geofences/
        POST /api/admin/geofences/
      summary: (관리자) 지오펜스 생성
      tags:
      - Admin — Geofence
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoFenceAdminRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GeoFenceAdminRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GeoFenceAdminRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoFenceAdmin'
          description: ''
        '400':
          description: 검증 실패
...
  /api/geofences/:
    get:
      operationId: geofences_list
      summary: 지오펜스 목록
      parameters:
      - in: query
        name: facility_id
        schema:
          type: integer
        description: 공장 ID 필터
      tags:
      - Geofence
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeoFence'
          description: ''
    post:
      operationId: geofences_create
      description: '다각형 좌표(`polygon`)와 위험도(`risk_level`: normal/warning/danger)로 생성.'
      summary: 지오펜스 생성
      tags:
      - Geofence
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoFenceRequest'
            examples:
              위험구역(사각형)생성:
                value:
                  facility: 1
                  name: A구역 — 압연기 주변
                  risk_level: danger
                  polygon:
                  - - 100
                    - 80
                  - - 300
                    - 80
                  - - 300
                    - 220
                  - - 100
                    - 220
                  description: 압연기 작동 중 접근 금지 구역
                summary: 위험구역 (사각형) 생성
                description: A공장 압연기 주변 4점 polygon
              주의구역(다각형)생성:
                value:
                  facility: 1
                  name: C구역 — 컨베이어
                  risk_level: warning
                  polygon:
                  - - 600
                    - 100
                  - - 900
                    - 100
                  - - 900
                    - 250
                  - - 750
                    - 350
                  - - 600
                    - 250
                  description: 컨베이어 작업 중 주의 필요
                summary: 주의구역 (다각형) 생성
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GeoFenceRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GeoFenceRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoFence'
          description: ''
...
components:
  schemas:
...
    GeoFence:
      type: object
      description: GeoFence 직렬화 — 상세 조회용
      properties:
        id:
          type: integer
          readOnly: true
        facility:
          type: integer
        name:
          type: string
          title: 구역 이름
          maxLength: 50
        polygon:
          title: 꼭짓점 좌표 배열
        risk_level:
          allOf:
          - $ref: '#/components/schemas/RiskLevel1beEnum'
          title: 위험도
        description:
          type: string
          title: 설명
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - facility
      - id
      - name
      - polygon
      - updated_at
    GeoFenceAdmin:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: 구역 이름
          maxLength: 50
        facility:
          type: integer
        facility_name:
          type: string
          readOnly: true
        risk_level:
          allOf:
          - $ref: '#/components/schemas/RiskLevel1beEnum'
          title: 위험도
        polygon:
          title: 꼭짓점 좌표 배열
        description:
          type: string
          title: 설명
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        worker_count:
          type: string
          readOnly: true
      required:
      - created_at
      - facility
      - facility_name
      - id
      - name
      - polygon
      - worker_count
    GeoFenceAdminListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/GeoFenceAdmin'
        total:
          type: integer
      required:
      - results
      - total
    GeoFenceAdminRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          title: 구역 이름
          maxLength: 50
        facility:
          type: integer
        risk_level:
          allOf:
          - $ref: '#/components/schemas/RiskLevel1beEnum'
          title: 위험도
        polygon:
          title: 꼭짓점 좌표 배열
        description:
          type: string
          title: 설명
        is_active:
          type: boolean
      required:
      - facility
      - name
      - polygon
    GeoFenceRequest:
      type: object
      description: GeoFence 직렬화 — 상세 조회용
      properties:
        facility:
          type: integer
        name:
          type: string
          minLength: 1
          title: 구역 이름
          maxLength: 50
        polygon:
          title: 꼭짓점 좌표 배열
        risk_level:
          allOf:
          - $ref: '#/components/schemas/RiskLevel1beEnum'
          title: 위험도
        description:
          type: string
          title: 설명
        is_active:
          type: boolean
      required:
      - facility
      - name
      - polygon
...
    RiskLevel1beEnum:
      enum:
      - normal
      - warning
      - danger
      type: string
      description: |-
        * `normal` - 정상
        * `warning` - 주의
        * `danger` - 위험
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT